requirements.txt
# grpcio==1.26.0
beautifulsoup4>=4.6.0
requests>=2.22.0
tqdm>=4.32.2
numpy>=1.18.0
pandas>=0.24.2
biopython>=1.73
tensorboard==2.1.0
h5py>=2.9.0
future>=0.18.2
openpyxl>=3.0.3
dask==2.21.0
distributed==2.21.0
dask-jobqueue==0.7.1
bokeh>=0.13.0
scikit-learn
matplotlib>=3.3.3
torch==1.7.1
torchvision==0.8.2
yapf
seaborn
einops==0.3.2
